+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 ### Modified
 
 - Completed the example so that it include all possible input formats.
+- Rename attribute `length` of a vehicle to `headway` for clarity (`length` is set as an alias for
+  backward compatibility).
 
 ### Fixed